requirements.txt
discord.py
python-dotenv
Pillow
giphy_client
requests